(HP-506) Added created env-config.js to HTML.
Added created test-env-config to test config.

Changed "process.env" usages to "window._env_"

Had to add an exception to eslint rules to allow "_env_"

config.ts uses the _env_ -variables when it is imported, which caused problems with tests. "window._env_" is not available at that time.

Fixed that in setupTests by mocking the config. Mocking required  importing window._env_  first and then return the config.

env variable overriding in tests had to be replaced and "mocked-env" package became obsolete